pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[project]
authors = [
  {name = "Michael A. Smith", email = "michael@smith-li.com"}
]
dependencies = [
  "djangorestframework (>=3.15.2,<4.0.0)",
  "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
  "boto3 (>=1.37.15,<2.0.0)",
  "django-extensions (>=3.2.3,<4.0.0)",
  "drf-spectacular (>=0.28.0,<0.29.0)"
]
description = "Managed Object Store by SmiTTY"
name = "moss"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[tool.autoflake]
expand-star-imports = true
remove-duplicate-keys = true
remove-unused-variables = true

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.vulture]
min_confidence = 100
paths = ["."]
